+++ src/eca/config.clj
@@ -19,7 +19,7 @@
   {:openaiApiKey nil
    :anthropicApiKey nil
    :rules []
-   :builtInTools {:filesystem {:enabled true}}
+   :nativeTools {:filesystem {:enabled true}}
    :mcpTimeoutSeconds 60
    :mcpServers {}
    :ollama {:host "http://localhost"