+++ src/eca/db.clj
@@ -9,17 +9,17 @@
    :chats {}
    :chat-behaviors ["agent" "chat"]
    :chat-default-behavior "chat"
-   :models {"o4-mini" {:mcp-tools true
+   :models {"o4-mini" {:tools true
                        :web-search false}
-            "o3" {:mcp-tools true
+            "o3" {:tools true
                   :web-search false}
-            "gpt-4.1" {:mcp-tools true
+            "gpt-4.1" {:tools true
                        :web-search true}
-            "claude-sonnet-4-0" {:mcp-tools true
+            "claude-sonnet-4-0" {:tools true
                                  :web-search true}
-            "claude-opus-4-0" {:mcp-tools true
+            "claude-opus-4-0" {:tools true
                                :web-search true}
-            "claude-3-5-haiku-latest" {:mcp-tools true
+            "claude-3-5-haiku-latest" {:tools true
                                        :web-search true}} ;; + ollama local models
    :default-model "o4-mini" ;; unless a ollama model is running.
    :mcp-clients {}})